+++ agents/__init__.py
@@ -1,5 +1,5 @@
 __all__ = ["GodAgent", "PopulationAgent", "WizardAgent", "EnhancedJudgeAgent"]
-from .god_agent import GodAgent
-from .population_agent import PopulationAgent
-from .wizard_agent import WizardAgent
-from .judge_agent import EnhancedJudgeAgent
+from agents.god_agent import GodAgent
+from agents.population_agent import PopulationAgent
+from agents.wizard_agent import WizardAgent
+from agents.judge_agent import EnhancedJudgeAgent
